moved ListProbes API view to separate file
