+++ poem/Poem/api/internal_views/probes.py
@@ -1,0 +1,43 @@
+from Poem.api import serializers
+from Poem.api.views import NotFound
+from Poem.poem_super_admin.models import Probe, ExtRevision
+
+from rest_framework.authentication import SessionAuthentication
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+
+class ListProbes(APIView):
+    authentication_classes = (SessionAuthentication,)
+
+    def get(self, request, name=None):
+        if name:
+            try:
+                probe = Probe.objects.get(name=name)
+                serializer = serializers.ProbeSerializer(probe)
+
+                return Response(serializer.data)
+
+            except Probe.DoesNotExist:
+                raise NotFound(status=404, detail='Probe not found')
+
+        else:
+            probes = Probe.objects.all()
+
+            results = []
+            for probe in probes:
+                # number of probe revisions
+                nv = ExtRevision.objects.filter(probeid=probe.id).count()
+                results.append(
+                    dict(
+                        name=probe.name,
+                        version=probe.version,
+                        docurl=probe.docurl,
+                        description=probe.description,
+                        comment=probe.comment,
+                        repository=probe.repository,
+                        nv=nv
+                    )
+                )
+
+            return Response(results)
